Check out my new blog on Zen of debugging in Python
Exploring the different ways to handle debugging in python. In summary, we’ve identified the below few points as the Zen of debugging in Python which may sometimes also extend to other languages, however our focus remains Python. Please note, this list is not exhaustive and is subject to change with suggestions.

#Python #ComputerScience #SoftwareDevelopment #Tech

https://thetechcruise.com/zen-of-debugging-python